requirements.txt
htmlmin==0.1.12
Jinja2==2.11.3
markdown2==2.4.0
MarkupSafe==1.1.1
pathtools==0.1.2
python-decouple==3.3
watchdog==0.10.2
